Implement paste reports
