+++ web/index.html
@@ -12,6 +12,17 @@
 
 <body>
     <div id="spinner-container" class="hidden"><div class="spinner"></div></div>
+    <div id="btn_report" class="hidden" title="Report paste">
+        <svg xmlns="http://www.w3.org/2000/svg" class="icon icon-tabler icon-tabler-flag" width="40"
+            height="40" viewBox="0 0 24 24" stroke-width="1.5" stroke="#bebebe" fill="none"
+            stroke-linecap="round" stroke-linejoin="round">
+            <path stroke="none" d="M0 0h24v24H0z" fill="none"/>
+            <line x1="5" y1="5" x2="5" y2="21" />
+            <line x1="19" y1="5" x2="19" y2="14" />
+            <path d="M5 5a5 5 0 0 1 7 0a5 5 0 0 0 7 0" />
+            <path d="M5 14a5 5 0 0 1 7 0a5 5 0 0 0 7 0" />
+        </svg>
+    </div>
     <div class="navigation">
         <div class="buttons" id="buttons_default">
             <button class="button" id="btn_new" title="Create new paste (Ctrl + Q)">
